@@ -10,15 +10,14 @@ packages = [
 ]
 
 [tool.poetry.dependencies]
-python = "^3.7"
-selenium = "^3.141.0"
+python = "^3.10"
+selenium = "^4.23.0"
 wget = "^3.2"
-requests = "^2.25.1"
+requests = "^2.32.3"
 
 [tool.poetry.dev-dependencies]
-autopep8 = "^1.5.7"
-flake8 = "^3.9.2"
-unittest = "^0.0"
+autopep8 = "^2.3.1"
+flake8 = "^7.1.0"
 
 [build-system]
 requires = ["poetry-core>=1.0.0", "setuptools >= 40.6.0", "wheel"]
