@@ -4,6 +4,7 @@ import webdriver_session.utils.chrome as utils
 import os
 from webdriver_session import ChromeSession
 from selenium import webdriver
+from selenium.webdriver.common.by import By
 import tracemalloc
 from time import sleep
 
@@ -13,12 +14,12 @@ tracemalloc.start()
 # python -m unittest tests.chrome.TestChromeUtils
 class TestChromeUtils(TestCase):
 
-    chrome_version = '93'
+    chrome_version = '127'
 
     # python -m unittest tests.chrome.TestChromeUtils.test_get_chrome_version
     def test_get_chrome_version(self):
         actual = utils.get_chrome_version()
-        self.assertRegex(self.chrome_version, '\d*')
+        self.assertRegex(self.chrome_version, r'\d*')
         self.assertTrue(self.__class__.chrome_version == actual)
 
     # python -m unittest tests.chrome.TestChromeUtils.test_download_chromedriver
@@ -28,7 +29,7 @@ class TestChromeUtils(TestCase):
 
     # python -m unittest tests.chrome.TestChromeUtils.test_get_chromedriver_version
     def test_get_chromedriver_version(self):
-        self.assertRegex(utils.get_chromedriver_version(os.getcwd()), '\d*')
+        self.assertRegex(utils.get_chromedriver_version(os.getcwd()), r'\d*')
         self.assertEqual(utils.get_chromedriver_version(
             os.getcwd()), self.__class__.chrome_version)
 
@@ -57,8 +58,8 @@ class TestChromeSession(TestCase):
         session = ChromeSession(download_path=os.getcwd())
         browser = session.get_browser()
         browser.get('http://speedtest.tele2.net/')
-        browser.find_element_by_link_text('1MB').click()
-        sleep(2)
+        browser.find_element(By.LINK_TEXT, '1MB').click()
+        sleep(3)
         self.assertTrue(os.path.isfile('1MB.zip'))
 
     # python -m unittest tests.chrome.TestChromeSession.test_add_pref
